refactor: move expressions module to chap_core/utils
